refactor: lombok 추가, 패키지 변경
